Share overlay init between visual-renderer.js and preview.js (#167)
Both the game (visual-renderer) and the preview harness repeated the same
overlay bring-up: getElementById("overlay-layer") → mountOverlays →
onViewport(reposition) → mountReticle → setReticleOverlay → onViewport. New
overlays had to be wired in both places.

Extract initializeGraphOverlays(layer?) into js/ui/overlays/index.js, which
mounts the registry overlays + the selection reticle, registers the reticle
with graph.js, and wires both to re-anchor on pan/zoom — returning
{ overlays, reticle }. Both entry points now call it, so future overlay
additions work in game + preview automatically. Behavior-preserving.
